fixing UI to domain name in menu. moved CORS perm decorator into lib/cors.py from app.py. removed test program. cleaned up on IDE
