+++ lib/cors.py
@@ -7,12 +7,15 @@ import json
 import base64
 import hmac
 import hashlib
+from functools import update_wrapper
 from datetime import datetime as dt
 from datetime import timedelta as td
 
 from flask import jsonify
+from flask import make_response
+from flask import current_app
+from flask import request
 from tabulate import tabulate
-
 
 def starts_with_branch(branch):
     """
@@ -71,3 +74,55 @@ class PolicySigner(object):
         return json dictionary so get request gets data filled because jsonify puts the right mime type on return
         """
         return jsonify(self.payload)
+
+
+def crossdomain(origin=None, methods=None, headers=None,
+                max_age=21600, attach_to_all=True,
+                automatic_options=True):
+    """
+    decorator that adjusts CORS permissions of controllers
+    :param origin:
+    :param methods:
+    :param headers:
+    :param max_age:
+    :param attach_to_all:
+    :param automatic_options:
+    :return:
+    """
+    if methods is not None:
+        methods = ', '.join(sorted(x.upper() for x in methods))
+    if headers is not None and not isinstance(headers, basestring):
+        headers = ', '.join(x.upper() for x in headers)
+    if not isinstance(origin, basestring):
+        origin = ', '.join(origin)
+    if isinstance(max_age, td):
+        max_age = max_age.total_seconds()
+
+    def get_methods():
+        if methods is not None:
+            return methods
+
+        options_resp = current_app.make_default_options_response()
+        return options_resp.headers['allow']
+
+    def decorator(f):
+        def wrapped_function(*args, **kwargs):
+            if automatic_options and request.method == 'OPTIONS':
+                resp = current_app.make_default_options_response()
+            else:
+                resp = make_response(f(*args, **kwargs))
+            if not attach_to_all and request.method != 'OPTIONS':
+                return resp
+
+            h = resp.headers
+
+            h['Access-Control-Allow-Origin'] = origin
+            h['Access-Control-Allow-Methods'] = get_methods()
+            h['Access-Control-Max-Age'] = str(max_age)
+            if headers is not None:
+                h['Access-Control-Allow-Headers'] = headers
+            return resp
+
+        f.provide_automatic_options = False
+        return update_wrapper(wrapped_function, f)
+    return decorator
