+++ templates/base.html
@@ -32,7 +32,7 @@
         </div>
         <div class="navbar-collapse collapse">
             <ul class="nav navbar-nav">
-                <li><a href="http://54.202.100.113:9000/#!/list/">List Files</a></li>
+                <li><a href="http://angular-s3.sfblur.com/#!/list/">List Files</a></li>
             </ul>
             <ul class="nav navbar-nav">
                 <li><a href="/about">About</a></li>
